requirements.txt
torch
torchvision
pillow
pandas
numpy
scikit-learn
matplotlib
seaborn
tqdm
streamlit
